Add trigger command functionality, and improve code structure.
